move MockAuthProvider and MockUrlSession into their own files
