+++ Tests/PlayolaPlayerTests/ListeningSessionTests.swift
@@ -8,61 +8,6 @@ import Testing
 import Foundation
 @testable import PlayolaPlayer
 
-// Mock auth provider for testing
-class MockAuthProvider: PlayolaAuthenticationProvider {
-    var currentToken: String?
-    var refreshedToken: String?
-    var refreshCallCount = 0
-    
-    init(currentToken: String? = nil, refreshedToken: String? = nil) {
-        self.currentToken = currentToken
-        self.refreshedToken = refreshedToken
-    }
-    
-    func getCurrentToken() async -> String? {
-        return currentToken
-    }
-    
-    func refreshToken() async -> String? {
-        refreshCallCount += 1
-        return refreshedToken
-    }
-}
-
-// Mock URLSession for testing
-class MockURLSession: URLSessionProtocol {
-    var responses: [(Data, URLResponse)] = []
-    var requestCallCount = 0
-    
-    func data(for request: URLRequest) async throws -> (Data, URLResponse) {
-        requestCallCount += 1
-        
-        if responses.isEmpty {
-            // Default successful response
-            let response = HTTPURLResponse(
-                url: request.url!,
-                statusCode: 200,
-                httpVersion: nil,
-                headerFields: nil
-            )!
-            return (Data(), response)
-        }
-        
-        // Return the first response and remove it (FIFO)
-        return responses.removeFirst()
-    }
-    
-    func addResponse(data: Data = Data(), statusCode: Int, url: URL? = nil) {
-        let response = HTTPURLResponse(
-            url: url ?? URL(string: "https://test.com")!,
-            statusCode: statusCode,
-            httpVersion: nil,
-            headerFields: nil
-        )!
-        responses.append((data, response))
-    }
-}
-
 @MainActor
 struct ListeningSessionTests {
